exp-client/requirements.txt
kivy>=2.3.0
kivymd==1.2.0
kivy-garden==0.1.5
kivy_garden.matplotlib

requests==2.31.0

SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
pg8000==1.30.5

PyJWT==2.10.1
cryptography==42.0.5
python-dotenv==1.0.0

matplotlib==3.8.4
pytest==8.1.1
setuptools==68.0.0